Allow setup.py/pip to run query commands without importing numpy.
pip downloads and queries packages before building them. If you are
running in a clean environment (e.g. a virtualenv) then numpy may not be
installed, and the query command (egg_info) could fail saying that numpy
cannot be imported. Solution is to check whether command is a query
and only import numpy if it is not. Idea taken from similar commit in
netcdf-python:
https://github.com/Unidata/netcdf4-python/commit/18f21a2ae2c21531f7d3bf8610cbc27cf2df2381
Should allow pip to install basemap from a requirements.txt file.
Bonus, add requirements for numpy>=1.2.1 and matplotlib>=1.0.0,
both taken from the basemap docs.