makes netCDF4 optional for examples/run_all.py
If you do not have netCDF4 installed, some tests will give ImportErrors.

This patch makes netCDF4 optional for run_all.py.  It imports netCDF4, if that package is not available, it removes tests that require netCDF4.

There may be some of the tests in the lines above the try-statement that could be put below the except-statement.  I did not look at that aspect.