add package_versions() and option to print info in test.py
package_versions() is a new function that returns a namedtuple of
version information of the dependent packages.  It should also be useful
to see if Travis is running the correct package versions.  I made this
with Issue #235 in mind.  A --verbose option was added to test.py to
display this version information.  Additionally, this new function could
serve a use in debugging future issues.