Multiple install environments
Downloads and compiles external version of libgeos from github.  Installs version that are "latest and greatest", "middle of the road", "minimum requirements", and "current development" environment.

The "middle of the road" Python 3.4.  Gets everything done in less than 3 minutes.

The "current development" environment is allowed to fail.  It is more of a diagnostic to address future possible incompatibilities.

For the "minimum requirements" environment, this took a bit of testing.  The earliest versions that I could get to work on Python 2.6 and 3.3 were matplotlib 1.2.0 and numpy 1.7.0.  Python 2.6 did work with earlier versions of numpy.  If desired, a travis test matrix could be built to test out all the possibilities to truly find the minimum.  See config for details.

Total build time is similar to before this, but this tests the boundaries a bit more.

There is more work that could be done, but I think this is a completely better system from the current.