Update __init__.py
Starting line 3960 within method drawlsmask I propose changing the use of numpy.concatenate to use np.hstack to achieve the stacking of the lsmask longitude array with itself. This is due to changes in numpy concatenate in NumPy version 1.10.0+ in which an IndexError will be raised if the axis of concatenation does not exist. Numpy.hstack will function the same way and persists in newer versions of NumPy.