Update test.py
Adds test of drawlsmask method to catch changes to NumPy functionality, which have broken drawlsmask recently. Recent changes to NumPy concatenate disallowed the concatenation of arrays on axes that don't exist. An IndexError is now raised when this is attempted. To test if drawlsmask functionality is working, the proposed addition is the TestLSMask class, which will fail the test if drawlsmask fails.