Add diagnostics functions: proj4 ver & Inv. Hammer
Refactored package_version function to move proj4 version checking into proj4_version function.  check_proj_inv_hammer function added to check if the inverse of the hammer projection is available (which should be in the yet to be released proj.4 version 4.9.3).

check_proj_inv_hammer() and proj4_version() are intended to end up in setup.py, but may provide some use in the main basemap library.  Due to that reason, the imports are mostly self-contained for the function.