Interpolate long line segments in lon,lat space.
New GSHHS versions simplify long national/state borders that are
perfectly linear in geographic space.  However, basemap assumes that
these lines are already subsampled, and does the interpolation in
between the two points in geographic space.  This leads to extreme
errors for long borders with only two points (e.g. the US/Canada
border). Old GSHHS versions stored interpolated versions, so this is an
issue that didn't come up in the past.

To get around this, we'll interpolate long linear segments and store the
interpolated versions.