Remove incompatibility with mpl v3.x
In mpl v3.x, cbook.is_scalar has been removed.  It is used in only
one place in basemap, and in that location it is not actually doing
anything useful.

I made a small reorganization and cleanup in that location in the
interp function to fix another bug in which masked locations were
being inadvertently unmasked without being filled with the mask
value specified in the kwarg.