Added checks and look up options for epsg file.
Since version 2.0, pyproj comes with an internal data directory that is used in favor of the one set by the PROJ_LIB environment variable (or, on Linux, of the default proj folder set by PROJ.4). It appears this internal directory sometimes does not contain the epsg file (at least on Linux), which may however be found by looking up those other locations.

This commit therefore adds attempts to look up for the epsg file when it cannot be found in pyproj's internal datadir.