Rewrite version manually in setup file
With the current structure inside the package init file, it is not
possible to rely on the package `__version__` value. If we do it,
we will get an ImportError because at setup time we still do not
have `geoslib` compiled or we might not have `pyproj` in our
environment.

At some point this can be restructured so that we can import
`__version__` like this in the future.