Downgrade numpy build requirement from 1.21.4 to 1.21.1
In principle, it is possible to use NumPy up to version 1.21.4
when building for Python >= 3.7, but the problem is that NumPy
versions from 1.21.2 to 1.21.6 add an explicit upper pin '<3.11'
which prevents `basemap` from being built successfully with the
recently released Python 3.11.

Since the goal here is to build the wheels with an old-enough
NumPy, we try here by downgrading NumPy at the build stage from
1.21.4 to 1.21.1, since NumPy 1.21.1 did not have this '<3.11'.

Note that we cannot go below NumPy 1.21.0 because then the MacOS
builds will fail due to a NumPy build bug.