Fixed plotting issue with hammer map data
Fixed in issue that the input data clipping causes when warping an existing map to a hammer map. The clipping of the data confuses the bluemarble method when it tries to draw the background. Removing the call the bluemarble in logic to warp to a hammer map fixes the plotting